world 注册服务 Gate 校验 login、转发 GameServer Game 服务

+++ server/game/internal/config.go
@@ -2,11 +2,11 @@ package internal
 
 var (
 	Key      = "FK44r1uPz9x7z8Ph75LIEI7nhp6H40R8"
-	serverID = 1
+	serverID = int32(1)
 	wsAddr   = "127.0.0.1:14561"
 )
 
-func SetServerID(id int) {
+func SetServerID(id int32) {
 	serverID = id
 }
